requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
openai==1.54.0
sounddevice==0.4.7
soundfile==0.12.1
aiofiles==23.1.0
python-multipart==0.0.6
requests==2.31.0
assemblyai==0.46.0
SQLAlchemy==2.0.23